Regras de negócio

+++ src/tasks/tasks.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable, UnprocessableEntityException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Task } from './task.entity';
@@ -10,7 +10,7 @@ export class TasksService {
   constructor(
     @InjectRepository(Task)
     private repo: Repository<Task>,
-  ) {}
+  ) { }
 
   findAll(): Promise<Task[]> {
     return this.repo.find();
@@ -22,16 +22,31 @@ export class TasksService {
 
   async create(dto: CreateTaskDto): Promise<Task> {
     const id = dto.id ?? uuidv4();
+    this.validate(dto);
     const entity = this.repo.create({ ...dto, id, done: dto.done ?? false });
     return this.repo.save(entity);
   }
 
+  validate(dto: Partial<CreateTaskDto>) {
+    if (dto.name.includes('Task')) {
+      throw new BadRequestException('O campo não deve conter o texto Task.');
+    }
+    if (dto.done === false && dto.completedDate) {
+      throw new BadRequestException('Não deve existir data de conclusão da tarefa para tarefas ativas. ')
+    }
+  }
+
   async update(id: string, dto: Partial<CreateTaskDto>) {
+    this.validate(dto);
     await this.repo.update({ id }, dto as any);
     return this.findOne(id);
   }
 
-  remove(id: string) {
+  async remove(id: string) {
+    const task = await this.repo.findOneBy({ id });
+    if(!task.done) {
+      throw new UnprocessableEntityException('Não é possível excluir uma tarefa ativa! ');
+    }
     return this.repo.delete({ id });
   }
 }
